[FIX] Under MS Windows, Point, Analog and Wrench object are not initialized correctly. [UPD] New constructor for Point, Analog and Wrench class with null matrices. The use of this constructor must be followed by the method SetFrameNumber() of each class.
git-svn-id: https://b-tk.googlecode.com/svn/BTK/trunk@272 93be6e1e-6fcc-11de-baec-69240fdc299a

+++ Code/Common/btkAcquisition.h
@@ -159,7 +159,7 @@ namespace btk
     BTK_COMMON_EXPORT void SetPointFrequency(double frequency);
     int GetAnalogFrameNumber() const {return this->m_PointFrameNumber * this->m_AnalogSampleNumberPerPointFrame;};    
     int GetNumberAnalogSamplePerFrame() const {return this->m_AnalogSampleNumberPerPointFrame;};
-    double GetAnalogFrequency() {return this->m_PointFrequency * this->m_AnalogSampleNumberPerPointFrame;};
+    double GetAnalogFrequency() {return this->m_PointFrequency * static_cast<double>(this->m_AnalogSampleNumberPerPointFrame);};
     AnalogResolution GetAnalogResolution() const {return this->m_AnalogResolution;};
     BTK_COMMON_EXPORT void SetAnalogResolution(AnalogResolution r);
